backend/requirements/base.txt
Django>=5.1
mysqlclient>=2.2
channels>=4.1
daphne>=4.1
channels-redis>=4.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.4
django-cors-headers>=4.4
django-environ>=0.11
celery[redis]>=5.4
httpx>=0.27
prometheus-client>=0.20
gunicorn>=23.0
uvicorn[standard]>=0.30
